Relocated helpers file so that it can be equally accessed by the files that need it.
